requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.1
asgiref==3.5.0
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==4.1.0
Brotli==1.0.9
matplotlib==3.7.1
certifi==2021.10.8
cffi==1.15.0
numerize==0.12
charset-normalizer==2.0.10
click==8.0.3
pyOpenSSL==16.2.0
azure-storage-blob==12.14.1
colorama==0.4.4
cycler==0.11.0
dash==2.3.1
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Django==4.0.4
django-cors-headers==3.12.0
django-crispy-forms==1.14.0
django-restframework==0.0.1
djangorestframework==3.13.1
dnspython==2.1.0
entrypoints==0.3
et-xmlfile==1.1.0
Flask==2.0.2
Flask-Compress==1.10.1
Flask-SeaSurf==0.3.1
fonttools==4.28.5
fqdn==1.5.1
humanize==4.0.0
hupper==1.10.3
idna==2.8
importlib-resources==5.4.0
ipykernel==6.6.1
ipympl==0.9.1
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonpointer==2.2
jsonschema==4.3.3
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kaleido==0.1.0
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.0.1
mistune==0.8.4
more-itertools==8.12.0
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.22.0
openapi-core==0.13.8
openapi-schema-validator==0.1.6
openapi-spec-validator==0.3.1
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
PasteDeploy==2.1.1
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.4.0
plaster==1.0
plaster-pastedeploy==0.7
plotly==5.5.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psutil==5.9.0
PuLP==2.6.0
pycparser==2.21
Pygments==2.11.1
pymongo==4.0.1
Pympler==1.0.1
pyparsing==3.0.6
pyramid==1.10.8
pyramid-openapi3==0.12
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
statsmodels==0.13.2
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.1.0
tornado==6.1
traitlets==5.1.1
translationstring==1.4
tzdata==2022.1
uri-template==1.1.0
urllib3==1.26.15
chardet==5.1.0
venusian==3.0.0
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
WebOb==1.8.7
Werkzeug==2.0.2
widgetsnbextension==3.6.0
zipp==3.7.0
zope.deprecation==4.4.0
zope.interface==5.4.0
SQLAlchemy==1.4.40
pyodbc==4.0.34